quantization/requirements.txt
auto-gptq 
optimum 
transformers 
accelerate
huggingface_hub
python-dotenv